ok, doing a swap with the files
